orays_linux_abi

Module constants

Linux UAPI constants that do not depend on an OrayS backend.

Constants

ACCESS_MODE_MASK

Mask of the valid access mode bits.

ACCESS_R_OK

R_OK mode bit for access/faccessat.

ACCESS_W_OK

W_OK mode bit for access/faccessat.

ACCESS_X_OK

X_OK mode bit for access/faccessat.

AF_UNIX_DOMAIN

AF_UNIX address family value.

AUX_CLOCK_TICKS

AT_CLKTCK auxiliary vector value: clock ticks per second (USER_HZ).

CHOWN_ID_UNCHANGED

Sentinel UID/GID value meaning “unchanged” for chown-family calls.

CLOSE_RANGE_CLOEXEC

CLOSE_RANGE_CLOEXEC flag for close_range.

CLOSE_RANGE_UNSHARE

CLOSE_RANGE_UNSHARE flag for close_range.

DEVFS_MAGIC

statfs magic of devfs.

EXT4_SUPER_MAGIC

statfs magic of the ext4 filesystem.

FD_SETSIZE

Maximum descriptor count of a Linux fd_set.

FILE_MODE_GROUP_EXECUTE

S_IXGRP group-execute mode bit.

FILE_MODE_PERMISSION_MASK

Permission bits mask of a file mode (07777).

FILE_MODE_SET_GID

S_ISGID set-group-ID mode bit.

FILE_MODE_SET_UID

S_ISUID set-user-ID mode bit.

FILE_MODE_STICKY

S_ISVTX sticky mode bit.

IOV_MAX

Maximum iovec count accepted by vectored IO syscalls.

IPPROTO_IP_LEVEL

IPPROTO_IP protocol level value.

IP_MCAST_JOIN_GROUP_OPT

IP_ADD_MEMBERSHIP-style multicast join option value used by Linux.

IP_MCAST_LEAVE_GROUP_OPT

IP_DROP_MEMBERSHIP-style multicast leave option value used by Linux.

IP_MULTICAST_LOOP_OPT

IP_MULTICAST_LOOP IP option.

IP_MULTICAST_TTL_OPT

IP_MULTICAST_TTL IP option.

IP_RECVERR_OPT

IP_RECVERR IP option.

IP_TOS_OPT

IP_TOS IP option.

IP_TTL_OPT

IP_TTL IP option.

KERNEL_SIGSET_BYTES

Size in bytes of the kernel sigset_t at the syscall boundary.

LINUX_EACCES

EACCES errno value.

LINUX_EAFNOSUPPORT

EAFNOSUPPORT errno value.

LINUX_ENOPROTOOPT

ENOPROTOOPT errno value.

LINUX_EOPNOTSUPP

EOPNOTSUPP errno value.

LINUX_EPROTONOSUPPORT

EPROTONOSUPPORT errno value.

LINUX_ESOCKTNOSUPPORT

ESOCKTNOSUPPORT errno value.

LINUX_PERSONALITY_QUERY

Sentinel persona value that queries instead of setting personality.

O_NOFOLLOW_FLAG

O_NOFOLLOW open flag.

O_PATH_FLAG

O_PATH open flag.

PERSONALITY_ADDR_COMPAT_LAYOUT

ADDR_COMPAT_LAYOUT personality flag.

PERSONALITY_ADDR_LIMIT_3GB

ADDR_LIMIT_3GB personality flag.

PERSONALITY_ADDR_LIMIT_32BIT

ADDR_LIMIT_32BIT personality flag.

PERSONALITY_ADDR_NO_RANDOMIZE

ADDR_NO_RANDOMIZE personality flag.

PERSONALITY_FDPIC_FUNCPTRS

FDPIC_FUNCPTRS personality flag.

PERSONALITY_KNOWN_FLAGS

Mask of every personality flag this boundary recognizes.

PERSONALITY_MAX_KNOWN_DOMAIN

Highest personality domain this boundary recognizes.

PERSONALITY_MMAP_PAGE_ZERO

MMAP_PAGE_ZERO personality flag.

PERSONALITY_PER_MASK

Mask of the personality domain bits.

PERSONALITY_READ_IMPLIES_EXEC

READ_IMPLIES_EXEC personality flag.

PERSONALITY_SHORT_INODE

SHORT_INODE personality flag.

PERSONALITY_STICKY_TIMEOUTS

STICKY_TIMEOUTS personality flag.

PERSONALITY_UNAME26

UNAME26 personality flag.

PERSONALITY_WHOLE_SECONDS

WHOLE_SECONDS personality flag.

PER_LINUX

PER_LINUX personality domain.

PIPEFS_MAGIC

statfs magic of pipefs.

PROC_SUPER_MAGIC

statfs magic of procfs.

RAMFS_MAGIC

statfs magic of ramfs.

RLIMIT_FSIZE_RESOURCE

RLIMIT_FSIZE resource number.

RLIMIT_NOFILE_RESOURCE

RLIMIT_NOFILE resource number.

RLIMIT_STACK_RESOURCE

RLIMIT_STACK resource number.

RTC_RD_TIME

RTC_RD_TIME ioctl request code.

SA_NODEFER_FLAG

SA_NODEFER sigaction flag.

SEEK_DATA_WHENCE

SEEK_DATA whence value for lseek.

SEEK_HOLE_WHENCE

SEEK_HOLE whence value for lseek.

SIGABRT_NUM

SIGABRT signal number.

SIGALRM_NUM

SIGALRM signal number.

SIGCANCEL_NUM

SIGCANCEL signal number used by Linux NPTL.

SIGCHLD_NUM

SIGCHLD signal number.

SIGCONT_NUM

SIGCONT signal number.

SIGFPE_NUM

SIGFPE signal number.

SIGILL_NUM

SIGILL signal number.

SIGINT_NUM

SIGINT signal number.

SIGKILL_NUM

SIGKILL signal number.

SIGPIPE_NUM

SIGPIPE signal number.

SIGPROF_NUM

SIGPROF signal number.

SIGQUIT_NUM

SIGQUIT signal number.

SIGSEGV_NUM

SIGSEGV signal number.

SIGSTOP_NUM

SIGSTOP signal number.

SIGTERM_NUM

SIGTERM signal number.

SIGVTALRM_NUM

SIGVTALRM signal number.

SIG_BLOCK_HOW

SIG_BLOCK how value for rt_sigprocmask.

SIG_SETMASK_HOW

SIG_SETMASK how value for rt_sigprocmask.

SIG_UNBLOCK_HOW

SIG_UNBLOCK how value for rt_sigprocmask.

SI_TKILL_CODE

SI_TKILL si_code value identifying tkill/tgkill signals.

SOL_SOCKET_LEVEL

SOL_SOCKET socket option level.

SO_ACCEPTCONN_OPT

SO_ACCEPTCONN socket option.

SO_BROADCAST_OPT

SO_BROADCAST socket option.

SO_DEBUG_OPT

SO_DEBUG socket option.

SO_DOMAIN_OPT

SO_DOMAIN socket option.

SO_DONTROUTE_OPT

SO_DONTROUTE socket option.

SO_ERROR_OPT

SO_ERROR socket option.

SO_KEEPALIVE_OPT

SO_KEEPALIVE socket option.

SO_LINGER_OPT

SO_LINGER socket option.

SO_NO_CHECK_OPT

SO_NO_CHECK socket option.

SO_OOBINLINE_OPT

SO_OOBINLINE socket option.

SO_PASSCRED_OPT

SO_PASSCRED socket option.

SO_PEERCRED_OPT

SO_PEERCRED socket option.

SO_PRIORITY_OPT

SO_PRIORITY socket option.

SO_PROTOCOL_OPT

SO_PROTOCOL socket option.

SO_RCVBUFFORCE_OPT

SO_RCVBUFFORCE socket option.

SO_RCVBUF_OPT

SO_RCVBUF socket option.

SO_RCVTIMEO_OPT

SO_RCVTIMEO socket option.

SO_REUSEADDR_OPT

SO_REUSEADDR socket option.

SO_REUSEPORT_OPT

SO_REUSEPORT socket option.

SO_SNDBUFFORCE_OPT

SO_SNDBUFFORCE socket option.

SO_SNDBUF_OPT

SO_SNDBUF socket option.

SO_SNDTIMEO_OPT

SO_SNDTIMEO socket option.

SO_TYPE_OPT

SO_TYPE socket option.

SS_DISABLE

SS_DISABLE sigaltstack flag disabling the alternate signal stack.

ST_MODE_BLK

S_IFBLK file-type bits of st_mode.

ST_MODE_CHR

S_IFCHR file-type bits of st_mode.

ST_MODE_DIR

S_IFDIR file-type bits of st_mode.

ST_MODE_FIFO

S_IFIFO file-type bits of st_mode.

ST_MODE_FILE

S_IFREG file-type bits of st_mode.

ST_MODE_LNK

S_IFLNK file-type bits of st_mode.

ST_MODE_SOCKET

S_IFSOCK file-type bits of st_mode.

ST_MODE_TYPE_MASK

S_IFMT file-type mask of st_mode.

SYSFS_MAGIC

statfs magic of sysfs.

SYSV_IPC_CREAT

IPC_CREAT flag for System V IPC get calls.

SYSV_IPC_EXCL

IPC_EXCL flag for System V IPC get calls.

SYSV_IPC_INFO

IPC_INFO control command.

SYSV_IPC_PRIVATE

IPC_PRIVATE key value for System V IPC creation.

SYSV_IPC_RMID

IPC_RMID control command.

SYSV_IPC_SET

IPC_SET control command.

SYSV_IPC_STAT

IPC_STAT control command.

SYSV_SHM_EXEC

SHM_EXEC shmat flag.

SYSV_SHM_HUGETLB

SHM_HUGETLB shmget flag.

SYSV_SHM_INFO

SHM_INFO shmctl command.

SYSV_SHM_LOCK

SHM_LOCK shmctl command.

SYSV_SHM_LOCKED

SHM_LOCKED state bit reported in shm_perm.mode.

SYSV_SHM_RDONLY

SHM_RDONLY shmat flag.

SYSV_SHM_REMAP

SHM_REMAP shmat flag.

SYSV_SHM_RND

SHM_RND shmat flag.

SYSV_SHM_STAT

SHM_STAT shmctl command.

SYSV_SHM_STAT_ANY

SHM_STAT_ANY shmctl command.

SYSV_SHM_UNLOCK

SHM_UNLOCK shmctl command.

TCP_KEEPCNT_OPT

TCP_KEEPCNT TCP option.

TCP_KEEPIDLE_OPT

TCP_KEEPIDLE TCP option.

TCP_KEEPINTVL_OPT

TCP_KEEPINTVL TCP option.

TCP_MAXSEG_OPT

TCP_MAXSEG TCP option.

TCP_NODELAY_OPT

TCP_NODELAY TCP option.

TMPFS_MAGIC

statfs magic of tmpfs.